errores
arreglando algunos errores
